Remove dead code and extract static methods from ModelResourceExporter (#524)
- Remove isMoreRecentThan (both overloads) — unused date comparisons
- Remove POTENTIAL_MODEL_CLASS_DATA_OPTIONS field — dead state
- Remove getBestClassDataForJointList — never called
- Remove getExistingJointIdPairs — only caller was getBestClassDataForJointList
- Remove private getAccessorMethodName — shadowed by delegation
- Remove stale commented-out getBestClassDataForJointList call
- Remove stale example comment block and main() method
- Extract getExistingJointIds to ModelResourceJavaGenerator
- Extract getAccessorMethodsForResourceClass to ModelResourceJavaGenerator
- Extract getJointAccessCodeForClass to ModelResourceJavaGenerator
- Update test call sites to use ModelResourceJavaGenerator

ModelResourceExporter: 886 → 742 lines (under 750 target)
All 68 model-loading tests pass, checkstyle clean.
