+++ core/model-loading/src/main/java/org/lgna/story/resourceutilities/ModelResourceExporter.java
@@ -244,22 +244,6 @@ public class ModelResourceExporter {
     return this.exportGalleryResources;
   }
 
-  public static boolean isMoreRecentThan(Date dataDate, File file) {
-    if (dataDate == null) {
-      return false;
-    }
-    Date fileDate = new Date(file.lastModified());
-    boolean isNewer = dataDate.after(fileDate);
-    return isNewer;
-  }
-
-  public static boolean isMoreRecentThan(Date dataData, Date otherDate) {
-    if (dataData == null) {
-      return false;
-    }
-    return dataData.after(otherDate);
-  }
-
   public String getResourceName() {
     return this.resourceName;
   }
@@ -359,10 +343,6 @@ public class ModelResourceExporter {
 
   public void setJointMap(List<Tuple2<String, String>> jointList) {
     this.jointList = jointList;
-    //    if (this.classData == null)
-    //    {
-    //      this.classData = ModelResourceExporter.getBestClassDataForJointList(jointList);
-    //    }
   }
 
   public void addSubResource(ModelSubResourceExporter subResource) {
@@ -526,93 +506,6 @@ public class ModelResourceExporter {
     return this.subResources;
   }
 
-  private static List<ModelClassData> POTENTIAL_MODEL_CLASS_DATA_OPTIONS = null;
-
-  public static ModelClassData getBestClassDataForJointList(List<Tuple2<String, String>> jointList) {
-    if (POTENTIAL_MODEL_CLASS_DATA_OPTIONS == null) {
-      POTENTIAL_MODEL_CLASS_DATA_OPTIONS = new LinkedList<ModelClassData>();
-      Field[] dataFields = AliceResourceClassUtilities.getFieldsOfType(ModelClassData.class, ModelClassData.class);
-      for (Field f : dataFields) {
-        ModelClassData data = null;
-        try {
-          Object o = f.get(null);
-          if ((o != null) && (o instanceof ModelClassData modelClassData)) {
-            data = modelClassData;
-          }
-        } catch (Exception e) {
-        }
-        if (data != null) {
-          POTENTIAL_MODEL_CLASS_DATA_OPTIONS.add(data);
-        }
-      }
-    }
-
-    int highScore = Integer.MIN_VALUE;
-    ModelClassData bestFit = null;
-    for (ModelClassData mcd : POTENTIAL_MODEL_CLASS_DATA_OPTIONS) {
-      List<Tuple2<String, String>> modelDataJoints = getExistingJointIdPairs(mcd.superClass);
-      int score = -Math.abs(modelDataJoints.size() - jointList.size());
-      for (Tuple2<String, String> inputPair : jointList) {
-        for (Tuple2<String, String> testPair : modelDataJoints) {
-          if (inputPair.equals(testPair)) {
-            score++;
-            break;
-          }
-        }
-      }
-      if (score > highScore) {
-        highScore = score;
-        bestFit = mcd;
-      }
-    }
-    return bestFit;
-  }
-
-  static List<Tuple2<String, String>> getExistingJointIdPairs(Class<?> resourceClass) {
-    List<Tuple2<String, String>> ids = new LinkedList<Tuple2<String, String>>();
-    Field[] fields = resourceClass.getDeclaredFields();
-    for (Field f : fields) {
-      if (JointId.class.isAssignableFrom(f.getType())) {
-        String fieldName = f.getName();
-        String parentName = null;
-        JointId fieldData = null;
-        try {
-          Object o = f.get(null);
-          if ((o != null) && (o instanceof JointId id)) {
-            fieldData = id;
-          }
-        } catch (Exception e) {
-        }
-        if ((fieldData != null) && (fieldData.getParent() != null)) {
-          parentName = fieldData.getParent().toString();
-        }
-
-        ids.add(Tuple2.createInstance(fieldName, parentName));
-      }
-    }
-    Class<?>[] interfaces = resourceClass.getInterfaces();
-    for (Class<?> i : interfaces) {
-      ids.addAll(getExistingJointIdPairs(i));
-    }
-    return ids;
-  }
-
-  static List<String> getExistingJointIds(Class<?> resourceClass) {
-    List<String> ids = new LinkedList<String>();
-    Field[] fields = resourceClass.getDeclaredFields();
-    for (Field f : fields) {
-      if (JointId.class.isAssignableFrom(f.getType())) {
-        String fieldName = f.getName();
-        ids.add(fieldName);
-      }
-    }
-    Class<?>[] interfaces = resourceClass.getInterfaces();
-    for (Class<?> i : interfaces) {
-      ids.addAll(getExistingJointIds(i));
-    }
-    return ids;
-  }
-
   Field getJointRootsField(Class<?> cls) {
     if (cls == null) {
       return null;
@@ -650,17 +543,6 @@ public class ModelResourceExporter {
       }
     }
     return false;
-  }
-
-  public static String getAccessorMethodsForResourceClass(Class<? extends JointedModelResource> resourceClass) {
-    StringBuilder sb = new StringBuilder();
-    List<String> jointIds = getExistingJointIds(resourceClass);
-    for (String id : jointIds) {
-      sb.append("public Joint get" + AliceResourceClassUtilities.getAliceMethodNameForEnum(id) + "() {\n");
-      sb.append("\t return org.lgna.story.Joint.getJoint( this, " + resourceClass.getCanonicalName() + "." + id + ");\n");
-      sb.append("}\n");
-    }
-    return sb.toString();
   }
 
   private static String createResourceEnumName(ModelResourceExporter parentExporter, String modelName, String textureName) {
@@ -702,10 +584,6 @@ public class ModelResourceExporter {
   String getJointAccessMethodNameForArrayJoint(String jointName) {
     String arrayName = ModelResourceArrayUtilities.getArrayNameForJoint(jointName, null, null);
     return ModelResourceJavaGenerator.getAccessorMethodName(arrayName);
-  }
-
-  private String getAccessorMethodName(String arrayName) {
-    return "get" + AliceResourceUtilities.enumToCamelCase(arrayName);
   }
 
   //If a parent interface has declared an accessor for a given field, return true
@@ -861,26 +739,4 @@ public class ModelResourceExporter {
     }
   }
 
-  /*
-   * public Joint getRightWrist() {
-   * return org.lgna.story.Joint.getJoint( this, org.lgna.story.resources.BipedResource.RIGHT_WRIST );
-   * }
-   */
-
-  public static String getJointAccessCodeForClass(Class<?> resourceClass) {
-    List<String> ids = getExistingJointIds(resourceClass);
-    StringBuilder sb = new StringBuilder();
-    for (String id : ids) {
-      sb.append("public org.lgna.story.Joint get" + AliceResourceClassUtilities.getAliceMethodNameForEnum(id) + "() {\n");
-      sb.append("\treturn org.lgna.story.Joint.getJoint( this, " + resourceClass.getName() + "." + id + " );\n");
-      sb.append("}\n");
-    }
-
-    return sb.toString();
-
-  }
-
-  public static void main(String[] args) {
-    System.out.println(getJointAccessCodeForClass(BipedResource.class));
-  }
 }
